desloppify: split alignment visual rendering
